Watchlist/requirements.txt
fastapi==0.112.0
uvicorn==0.30.5
python-jose==3.3.0
sqlalchemy==2.0.32
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
httpx==0.27.2